+++ main_exercise/README.md
@@ -1,38 +1,53 @@
-# Exercise 4 - Express.js
+# Exercise 5 - Sequelize ORM
 
-You know what they say about #4: you'll remember it *for*ever (ha)
+It is not persisted until you persist it!
 
 ## In this section you will practice
 
-**Express** - One of the most popular server frameworks for node.js
-**APIs** - We've already worked with an external API (the Pokemon one) - now we're going to build our own!
-**CRUD** - Create, Read, Update, Delete - the four fundamental operations of any API. We aren't updating yet, but the principles are the same
+**Initializing Sequelize ORM** - Connect NodeJS application to your mysql DB using Sequelize ORM 
+
+**Sequelize models** - Use Sequelize models to execute queries on your DB
+
+**Migrations and seeds** - Manage DB changes using Sequelize migrations
 
 ## What you are going to build
 
-So far our whole app has been 100% local, and if you refreshed the page - that's it, all your todos were gone. No persistency. So sad.
+In the last exercise, you've added an ExpressJS server to your todo app, which allowed you to reload your todos from server. 
 
-But that changes today! We're going to create a server that will handle all our todo items, _and_ our Pokemon fetching, and finally take our first step in becoming real #fullstack developers #hashtag
+But what happens when you restart your server?! you guessed right, all todos disappeared :( 
 
-You can use your existing project (copy+paste then refactor), or use the boilerplate we've set up for you in this `ex4` directory
+In this exercise we will add a DB to our application that will hold all items' data. This will provide us a real persistent storage that would keep our data even if our server is down. 
+
+You can use your ex4 solution or use the boilerplate in this folder. 
+
+### Prerequisites:
+Following pre-requisites were covered in our last workshop. 
+- Download and install [docker](https://docs.docker.com/get-docker/)
+- Open console and pull latest mysql image: ```docker pull mysql/mysql-server ```
+- Run mysql container and initialize it with the proper user, password, db name and permissions: ```docker run -p 3306:3306 --name tododb -e MYSQL_ROOT_PASSWORD=password -e MYSQL_ROOT_HOST=% -e MYSQL_DATABASE=todo_db -d mysql/mysql-server```
+- Validate container is up: ```docker ps``` 
 
 ### The requirements:
 
-- [ ] Create your express backend (include separate `dist` and `server` folders)
-- [ ] Your `server.js` file should have all the express boilerplate and host your `dist` directory to any client that requests it (hint: you'll need to `.use` the `express.static` method)
-- [ ] Create an `api.js` file that acts as the 'controller' of your backend, handling all the routes (endpoints)
-- [ ] Create separate endpoints for (1) fetching all the todo items, (2) creating a new one, and (3) deleting an existing one (hint: don't forget `bodyParser`)
-- [ ] Move the pokemon fetching code to the backend - use `axios` instead of `fetch` for your requests
-- [ ] On app load (i.e. when a user enters the page) it should fetch all the todo items and render them
+- [ ] Install Sequelize and mysql driver. [Sequelize- Getting Started](https://sequelize.org/docs/v6/getting-started/)
+- [ ] Install Sequelize CLI. [Installing the CLI](https://sequelize.org/docs/v6/other-topics/migrations/)
+- [ ] Initialize Sequelize using `npx sequelize-cli init` inside 'src/server/db' folder 
+- [ ] Create Items table using [Sequelize migration](https://sequelize.org/docs/v6/other-topics/migrations/#creating-the-first-model-and-migration) - a new table with id and ItemName fields
+- [ ] Modify `item_manager.js`: remove items array and modify all item operations to use Item model
+- [ ] Create and run a separate migration for adding a `status` column (BOOLEAN) to Items table in your DB
+- [ ] Add checkbox to each item in UI to indicate its status (Done vs not)
+- [ ] Modify client and server code to support persistence of the new Item status 
 
-In terms of the front end, it will look the same as before:
-![](src/assets/hw-2.gif)
+Your todo app should have now an additional checkbox that marks the status of the item. Every change to the checkbox should be stored in our Items table under the status column (true or false)
 
-But now when you refresh the page **the data should still be there**
+Now, even if your server is down - all your items are stored. Once the server is up again - you should be able to see all items.
+
+Here is an example how it can look on the client side:
+![](../assets/hw-5.gif)
 
 ### Bonus
 
-- [ ] Create a [middleware](https://expressjs.com/en/guide/using-middleware.html) that makes a log each time a user accesses any of the routes (you can just do a `console.log`)
-- [ ] Handle server errors elegantly. Specifically, if anything goes wrong the user should see an error message (ideally, not an alert) with an explanation of what went wrong instead of crashing the page
-- [ ] Add a loder/spinner to the page that indicates the client is waiting for an async operation (e.g. a call to the server) to finish
-- [ ] Add simple caching to your server. If a user requests for the same pokemon ID three times in the same minute, for example, it should only make a request to the Pokemon API once. You can use a simple in-memory data structure for your cache
+- [ ] Add "Done" timestamp
+- [ ] Add index to the Items table (which columns compose the index?) 
+- [ ] Add server validation - create a new item only if not exists (Use transaction)
+- [ ] Add edit capabilities to an item.
